Update to webpack 5
