+++ package.json
@@ -18,10 +18,10 @@
   },
   "scripts": {
     "test": "npm run lint",
-    "build": "webpack --config webpack.prod.ts",
+    "build": "webpack --mode production",
     "download-example-input": "ts-node ./utils/download-example-input.ts",
     "api-docs": "typedoc",
-    "start": "webpack-dev-server --config webpack.dev.ts --content-base ./app",
+    "start": "webpack serve --mode development --content-base ./app",
     "chemiscope-dts": "dts-bundle-generator -o dist/chemiscope.d.ts dist/src/index.d.ts",
     "jsmol-widget-dts": "dts-bundle-generator -o dist/jsmol-widget.d.ts dist/src/structure/widget.d.ts",
     "merge-dts": "npm run chemiscope-dts && npm run jsmol-widget-dts",
@@ -34,10 +34,10 @@
     "@types/node": "^14.0.23",
     "@types/plotly.js": "^1.50.15",
     "@types/tmp": "^0.2.0",
-    "@types/webpack": "^4.41.21",
-    "@types/webpack-merge": "^4.1.5",
     "@typescript-eslint/eslint-plugin": "^4.5.0",
     "@typescript-eslint/parser": "^4.5.0",
+    "assert": "^2.0.0",
+    "buffer": "^6.0.2",
     "css-loader": "^5.0.0",
     "dts-bundle-generator": "^5.1.0",
     "eslint": "^7.5.0",
@@ -48,6 +48,7 @@
     "markdown-it": "^12.0.2",
     "plotly.js": "1.54",
     "prettier": "2.1.2",
+    "process": "^0.11.10",
     "style-loader": "^2.0.0",
     "tmp": "^0.2.1",
     "ts-loader": "^8.0.1",
@@ -55,9 +56,8 @@
     "typedoc": "^0.17.8",
     "typedoc-plugin-external-module-name": "^4",
     "typescript": "^4.0.3",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.0",
-    "webpack-merge": "^4.0"
+    "webpack": "^5.2.0",
+    "webpack-cli": "^4.1.0",
+    "webpack-dev-server": "^3.11.0"
   }
 }
