--- webpack.dev.ts
@@ -1,11 +1,0 @@
-import webpack from 'webpack';
-import merge from 'webpack-merge';
-
-import base from './webpack.base';
-
-const config: webpack.Configuration = merge(base, {
-    devtool: 'inline-source-map',
-    mode: 'development',
-});
-
-export default config;
